CC 5321 Float stock revert and module dependency adjustment

+++ Bundles/AvailabilityOfferConnector/src/Spryker/Zed/AvailabilityOfferConnector/Dependency/Facade/AvailabilityOfferConnectorToAvailabilityFacadeBridge.php
@@ -28,9 +28,9 @@ class AvailabilityOfferConnectorToAvailabilityFacadeBridge implements Availabili
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
      *
-     * @return float
+     * @return int
      */
-    public function calculateStockForProductWithStore($sku, StoreTransfer $storeTransfer): float
+    public function calculateStockForProductWithStore($sku, StoreTransfer $storeTransfer): int
     {
         return $this->availabilityFacade->calculateStockForProductWithStore($sku, $storeTransfer);
     }
